+++ src/api/map/getPharmacyDetail.ts
@@ -1,5 +1,5 @@
 import axios from "../axiosConfig";
-import { PharmacyDetailType } from "@pages/map/mapTypes";
+import { PharmacyDetailType } from "./types";
 
 export default async function getPharmacyDetail(hpid: string) {
 	const response = await axios.get<PharmacyDetailType>(`/maps/${hpid}`);
