+++ src/api/map/types.ts
@@ -1,14 +1,3 @@
-export interface Pharmacy {
-	lat: number;
-	lng: number;
-	name: string;
-	hpid: string;
-}
-
-export interface MapProps {
-	pharmacies: Pharmacy[];
-}
-
 export type hourListType = {
 	dayOfWeek: string;
 	startHour: string;
