+++ src/pages/map/map.types.ts
@@ -1,0 +1,10 @@
+export interface Pharmacy {
+	lat: number;
+	lng: number;
+	name: string;
+	hpid: string;
+}
+
+export interface MapProps {
+	pharmacies: Pharmacy[];
+}
